requirements.txt
Flask==2.3.2
torch==2.0.1
torchvision==0.15.2
diffusers==0.21.4
transformers==4.30.2
Pillow==10.0.0
python-dotenv==1.0.0
